--- africanus/dft/__init__.py
@@ -1,2 +1,0 @@
-
-from .direct_fourier_transform import dft
