+++ africanus/rime/__init__.py
@@ -1,0 +1,2 @@
+
+from .phase import phase_delay
